+++ src/extension/network/tcpSend.cpp
@@ -31,7 +31,7 @@ namespace NUClear {
             // Construct a packet to send
             network::DataPacket p;
             p.type = network::DATA;
-            p.length = (sizeof(network::DataPacket) - 1) - sizeof(network::PacketHeader) + emit.data.size();
+            p.length = uint32_t((sizeof(network::DataPacket) - 1) - sizeof(network::PacketHeader) + emit.data.size());
             p.packetId = ++packetIDSource;
             p.packetNo = 1;     // TCP packets always come in one packet
             p.packetCount = 1;  // TCP packets always come in one packet
